+++ robots-proj/make.nu
@@ -1,0 +1,1 @@
+mvn clean package
