+++ robots-proj/src/main/java/robots/gui/GameWindow.java
@@ -1,0 +1,19 @@
+package robots.gui;
+
+import java.awt.BorderLayout;
+
+import javax.swing.JInternalFrame;
+import javax.swing.JPanel;
+
+public class GameWindow extends JInternalFrame {
+    private final GameVisualizer m_visualizer;
+
+    public GameWindow() {
+        super("Игровое поле", true, true, true, true);
+        m_visualizer = new GameVisualizer();
+        JPanel panel = new JPanel(new BorderLayout());
+        panel.add(m_visualizer, BorderLayout.CENTER);
+        getContentPane().add(panel);
+        pack();
+    }
+}
