+++ robots-proj/src/main/java/robots/gui/RobotsProgram.java
@@ -1,0 +1,25 @@
+package robots.gui;
+
+import java.awt.Frame;
+import javax.swing.SwingUtilities;
+import javax.swing.UIManager;
+
+public class RobotsProgram {
+    public static void main(String[] args) {
+        //Locale.setDefault(Locale.JAPAN); //(new Locale("ru", "RU"));
+        try {
+            UIManager.setLookAndFeel("javax.swing.plaf.nimbus.NimbusLookAndFeel");
+            // UIManager.setLookAndFeel("javax.swing.plaf.metal.MetalLookAndFeel");
+            // UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+            // UIManager.setLookAndFeel(UIManager.getCrossPlatformLookAndFeelClassName());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        SwingUtilities.invokeLater(() -> {
+            MainApplicationFrame frame = new MainApplicationFrame();
+            frame.pack();
+            frame.setVisible(true);
+            frame.setExtendedState(Frame.MAXIMIZED_BOTH);
+        });
+    }
+}
