+++ robots-proj/src/main/java/robots/log/LogChangeListener.java
@@ -1,0 +1,5 @@
+package robots.log;
+
+public interface LogChangeListener {
+    public void onLogChanged();
+}
