+++ robots-proj/src/main/java/robots/log/LogWindowSource.java
@@ -1,0 +1,76 @@
+package robots.log;
+
+import java.util.ArrayList;
+import java.util.Collections;
+
+/**
+ * Что починить: 1. Этот класс порождает утечку ресурсов (связанные слушатели оказываются
+ * удерживаемыми в памяти) TODO
+ * 2. Этот класс хранит активные сообщения лога, но в такой реализации он их
+ * лишь накапливает. Надо же, чтобы количество сообщений в логе было ограничено величиной
+ * m_iQueueLength (т.е. реально нужна очередь сообщений ограниченного размера)
+ */
+public class LogWindowSource {
+    private int m_iQueueLength;
+
+    private ArrayList<LogEntry> m_messages;
+    private final ArrayList<LogChangeListener> m_listeners;
+    private volatile LogChangeListener[] m_activeListeners;
+
+    public LogWindowSource(int iQueueLength) {
+        m_iQueueLength = iQueueLength;
+        m_messages = new ArrayList<LogEntry>(iQueueLength);
+        m_listeners = new ArrayList<LogChangeListener>();
+    }
+
+    public void registerListener(LogChangeListener listener) {
+        synchronized (m_listeners) {
+            m_listeners.add(listener);
+            m_activeListeners = null;
+        }
+    }
+
+    public void unregisterListener(LogChangeListener listener) {
+        synchronized (m_listeners) {
+            m_listeners.remove(listener);
+            m_activeListeners = null;
+        }
+    }
+
+    public void append(LogLevel logLevel, String strMessage) {
+        LogEntry entry = new LogEntry(logLevel, strMessage);
+        if (size() == m_iQueueLength) {
+            m_messages.remove(0);
+        }
+        m_messages.add(entry);
+
+        LogChangeListener[] activeListeners = m_activeListeners;
+        if (null == activeListeners) {
+            synchronized (m_listeners) {
+                if (null == m_activeListeners) {
+                    activeListeners = m_listeners.toArray(new LogChangeListener[0]);
+                    m_activeListeners = activeListeners;
+                }
+            }
+        }
+        for (LogChangeListener listener : activeListeners) {
+            listener.onLogChanged();
+        }
+    }
+
+    public int size() {
+        return m_messages.size();
+    }
+
+    public Iterable<LogEntry> range(int startFrom, int count) {
+        if (startFrom < 0 || m_messages.size() <= startFrom) {
+            return Collections.emptyList();
+        }
+        int indexTo = Math.min(startFrom + count, m_messages.size());
+        return m_messages.subList(startFrom, indexTo);
+    }
+
+    public Iterable<LogEntry> all() {
+        return m_messages;
+    }
+}
