--- robots/src/log/LogChangeListener.java
@@ -1,6 +1,0 @@
-package log;
-
-public interface LogChangeListener
-{
-    public void onLogChanged(); 
-}
